requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Pillow==5.0.0
python-dateutil==2.6.1
selenium==3.8.1
six==1.11.0
